backend/requirements.txt
fastapi>=0.100.0
httpx>=0.24.1
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0 